HSM simple transition working, all resolutions are statics. Added also helper methods
User just need the type of the state to trigger a transition to it with the trigger_transition<CUSTOM_STATE_T>() method. The new ascend_state_hierarchy<...>() and drill_state_hierarchy<...>() methods  are parts of this new transition triggering system.

Also used by the strate transition system but not depend on it, I have added seek_state<CUSTOM_STATE_T>() constexpr static helper method allowing to know if a CUSTOM_STATE is present among all the sub(n)-states of a state. The second constexpr static helper method added is has_substates(), allowing to know if a state has any substates or not.